+++ ui/config.go
@@ -2,6 +2,7 @@ package ui
 
 import (
 	"github.com/Hecatoncheir/lazyrest/environment"
+	"github.com/Hecatoncheir/lazyrest/keymap"
 	parserhttp "github.com/Hecatoncheir/lazyrest/parser/http"
 	"github.com/Hecatoncheir/lazyrest/runner"
 )
@@ -11,4 +12,5 @@ type Config struct {
 	ParseOptions    parserhttp.ParseOptions
 	EnvironmentName string
 	Environment     environment.Config
+	Keybindings     *keymap.Bindings
 }