+++ ui/producer/parameters.go
@@ -1,6 +1,7 @@
 package producer
 
 import (
+	"github.com/Hecatoncheir/lazyrest/keymap"
 	"github.com/Hecatoncheir/lazyrest/runner"
 	"github.com/Hecatoncheir/lazyrest/ui/theme"
 	"github.com/rivo/tview"
@@ -13,4 +14,5 @@ type Parameters struct {
 	OnRunFinishedCallback func(runner.Response, error)
 	App                   *tview.Application
 	RunnerConfig          runner.Config
+	Keybindings           *keymap.Bindings
 }