+++ ui/producer/producer.go
@@ -2,6 +2,7 @@ package producer
 
 import (
 	"context"
+	"github.com/Hecatoncheir/lazyrest/keymap"
 	"github.com/Hecatoncheir/lazyrest/parser/http"
 	"github.com/Hecatoncheir/lazyrest/runner"
 	"github.com/Hecatoncheir/lazyrest/ui/theme"
@@ -37,6 +38,7 @@ type Producer struct {
 	searchQuery      string
 	runnerConfig     runner.Config
 	bodyViewMode     BodyViewMode
+	keybindings      *keymap.Bindings
 }
 
 func (widget *Producer) StartRun() (context.Context, uint64) {
@@ -87,11 +89,15 @@ func (widget *Producer) CancelActive() {
 }
 
 func (widget *Producer) Build(parameters Parameters) tview.Primitive {
+	if parameters.Keybindings == nil {
+		parameters.Keybindings = keymap.Default()
+	}
 	widget.onEscapeCallback = parameters.OnEscapeCallback
 	widget.onProgress = parameters.OnProgressCallback
 	widget.onRunFinished = parameters.OnRunFinishedCallback
 	widget.app = parameters.App
 	widget.runnerConfig = parameters.RunnerConfig
+	widget.keybindings = parameters.Keybindings
 	widget.bodyViewMode = BodyViewPretty
 	theme := parameters.Theme.Producer
 	widget.theme = theme