+++ ui/suite/on_input_callback.go
@@ -1,6 +1,7 @@
 package suite
 
 import (
+	"github.com/Hecatoncheir/lazyrest/keymap"
 	"github.com/Hecatoncheir/lazyrest/parser/http"
 
 	"github.com/gdamore/tcell/v2"
@@ -12,11 +13,13 @@ type OnRunCallbackType func(suite http.HttpSuite)
 
 func onInputCallback(widget *Suite) func(event *tcell.EventKey) *tcell.EventKey {
 	return func(event *tcell.EventKey) *tcell.EventKey {
-		switch event.Key() {
-		case tcell.KeyEsc:
+		switch {
+		case widget.keybindings.Matches(keymap.Back, event):
 			widget.onEscapeCallback()
-		case tcell.KeyEnter:
+			return nil
+		case widget.keybindings.Matches(keymap.Run, event):
 			widget.onRunCallback(widget.suite)
+			return nil
 		}
 		return event
 	}