+++ ui/suite/parameters.go
@@ -1,9 +1,13 @@
 package suite
 
-import "github.com/Hecatoncheir/lazyrest/ui/theme"
+import (
+	"github.com/Hecatoncheir/lazyrest/keymap"
+	"github.com/Hecatoncheir/lazyrest/ui/theme"
+)
 
 type Parameters struct {
 	Theme            theme.Theme
 	OnEscapeCallback OnEscapeCallbackType
 	OnRunCallback    OnRunCallbackType
+	Keybindings      *keymap.Bindings
 }