+++ ui/suites/parameters.go
@@ -1,6 +1,7 @@
 package suites
 
 import (
+	"github.com/Hecatoncheir/lazyrest/keymap"
 	"github.com/Hecatoncheir/lazyrest/parser/http"
 	"github.com/Hecatoncheir/lazyrest/ui/theme"
 )
@@ -10,4 +11,5 @@ type Parameters struct {
 	OnEscapeCallback          OnEscapeCallbackType
 	OnSuiteSelectCallbackType OnSuiteSelectCallbackType
 	ParseOptions              http.ParseOptions
+	Keybindings               *keymap.Bindings
 }