+++ ui/tree/parameters.go
@@ -1,6 +1,9 @@
 package tree
 
-import "github.com/Hecatoncheir/lazyrest/ui/theme"
+import (
+	"github.com/Hecatoncheir/lazyrest/keymap"
+	"github.com/Hecatoncheir/lazyrest/ui/theme"
+)
 
 type Parameters struct {
 	RootDirectoryPath    string
@@ -8,4 +11,5 @@ type Parameters struct {
 	Theme                theme.Theme
 	OnSelectFileCallback OnSelectFileCallbackType
 	OnReloadCallback     func()
+	Keybindings          *keymap.Bindings
 }